Refactor the code to render without hard-coding and clear input field
We wrap the for-loop function in a renderLeads() function we create and call it inside the Event handler for the button clicks. That is inputBtn.addEventListener(click, function saveLead(){}). This way the items are pushed into the myLeads array and rendered out after the user enters the leads in the input field and presses the save input button. we also clear the input field after each entry.